Codigo Agente

+++ web/src/app/@core/guards/auth-interceptos.ts
@@ -11,17 +11,19 @@ import { Observable, throwError } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 import { Router } from '@angular/router';
 import { OAuth2Service } from '../services/util/oauth2.service';
+import { AlertService } from '../services/util/alert.service';
 
 @Injectable()
 export class AuthInterceptor implements HttpInterceptor {
-  constructor(private router: Router, protected oauth2Service: OAuth2Service) {}
+  constructor(private router: Router, protected oauth2Service: OAuth2Service, protected alertService: AlertService) {}
 
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     return next.handle(req).pipe(
       catchError((error: HttpErrorResponse) => {
-        if (error.status === 401 || error.status === 403) {
+        if ((error.status === 401 || error.status === 403) && !req.url.includes('/auth')) {
             this.router.navigateByUrl("/auth");
             this.oauth2Service.logout().then((response: any) => {
+                this.alertService.showWarning("Seu token não é válido, favor fazer login novamente.");
             });
             sessionStorage.clear();
         }
